+++ src/main.ts
@@ -2,8 +2,10 @@ import {
   bootstrapApplication,
   provideProtractorTestingSupport,
 } from '@angular/platform-browser';
+import { provideRouter } from '@angular/router';
+import routeConfig from './app/routing/routes';
 import { RootComponent } from './app/components/root/root.component';
 
 bootstrapApplication(RootComponent, {
-  providers: [provideProtractorTestingSupport()],
+  providers: [provideProtractorTestingSupport(), provideRouter(routeConfig)],
 }).catch((err) => console.error(err));