+++ HISTORY.md
@@ -17,3 +17,4 @@
 1. [Add an interpolation to a component’s template](https://angular.io/tutorial/first-app/first-app-lesson-07)
 1. [Use `*ngFor` to list objects in component](https://angular.io/tutorial/first-app/first-app-lesson-08)
 1. [`ng generate service housing`](https://angular.io/tutorial/first-app/first-app-lesson-09)
+1. [Add routes to the application](https://angular.io/tutorial/first-app/first-app-lesson-10)